Allow exposing private method correctly
By replacing '_' by '-' in method names for subparsers 'private' methods
will end up having a name that is not appropriate for command line parsing.
Hence we remove the leading '_' before making the replace.

Also added a TODO to make the class decorator a class to pass it some arguments
such as a specific description - to avoid using a docstring that's usually not
for that - and a bool to expose or not private methods.