Use a special HelpAction to print help for subparsers as well
Top parser help `python script.py --help` only display the name of the sub-commands
and not their full help which often forces the user to first check the name
of sub-command and then type `python script.py sub-command --help` to have
the required message.

This HelpAction taken from http://stackoverflow.com/a/24122778/2003420
will do just that. It is then added as the action of the `--help` argument
of the top level argument parser.