Make it possible to parse and call the proper method in one line!
Using the parser we are able to retrieve the method and arguments to call it
with. If the __init__ method is decorated an instance of the decorated class
is created on the fly. Otherwise an instance can be passed.

This is pretty awesome because the user do NOT have to know anything about
argparse!!!

Tests:
python setup.py nosetests