Handle bool arguments with default value as flags
Bool argument should have an action="store_{true|false}", not a type and be
used as flags.

Note: if the default value is True we use store_false, and vice-versa, as an
action. Passing the flag to the command line will 'flip the switch'.

Tests:
Need to test for args without default value.
python setup.py nosetests