Introduce usage of annotations
The annotations from getfullargspec are passed to define the type of the
cli parser's arguments.

The annotations will provide the necessary typing to convert the cli
argument to the proper python type.

The annotations are taking precedence over the types provider to the
decorator since we want to only leverage annotations in the long run.

Tests:
- pytest test OK